add ability to debug multi run

+++ scripts/multi_run.bash
@@ -1,6 +1,7 @@
 #!/usr/bin/env bash
 
 set -euo pipefail
+OPTIND=1         # Reset in case getopts has been used previously in the shell.
 
 declare -r DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 
@@ -30,7 +31,25 @@ for ip in $(jq -rc '.[].NetAddr' "$PEERS_DIR/lachesis_data_dir/peers.json"); do
 done
 
 # Run multi lachesis
-GOMAXPROCS=$(($logicalCpuCount - 1)) "$BUILD_DIR/lachesis_$TARGET_OS" run --datadir "$BUILD_DIR/lachesis_data_dir" --store --listen="$node_addr":12000 --log=warn --heartbeat=5s -p "$node_addr":9000 --test --test_n=10 --test_delay=10
+
+declare debug=0
+while getopts "d" opt; do
+    case "$opt" in
+    d)  debug=1
+        ;;
+    esac
+done
+
+shift $((OPTIND-1))
+
+[ "${1:-}" = "--" ] && shift
+
+if [ "$debug" == 0 ]; then
+  GOMAXPROCS=$(($logicalCpuCount - 1)) "$BUILD_DIR/lachesis_$TARGET_OS" run --datadir "$BUILD_DIR/lachesis_data_dir" --store --listen="$node_addr":12000 --log=warn --heartbeat=5s -p "$node_addr":9000 --test --test_n=10 --test_delay=10
+  rm -rf "$BUILD_DIR/lachesis_data_dir/"
+else
+  GOMAXPROCS=$(($logicalCpuCount - 1)) dlv --listen=localhost:37555 --headless=true --api-version=2 --backend=default exec "$BUILD_DIR/lachesis_$TARGET_OS" -- run --datadir "$BUILD_DIR/lachesis_data_dir" --store --listen="$node_addr":12000 --log=warn --heartbeat=5s -p "$node_addr":9000 --test --test_n=10 --test_delay=10
+fi
+
 declare -i rc=$?
-rm -rf "$BUILD_DIR/lachesis_data_dir/"
 exit "$rc"
